more kpi

+++ src/data/KPI.sql
@@ -9,13 +9,12 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE TABLE [dbo].[KPI11](
---	ID bigint identity not null,
-	[IDClient] [bigint] NOT NULL,
+	[UserId] nvarchar(1000) not null,
+--	[IDClient] [bigint] NULL,
 	[IDAssVA] [bigint] NOT NULL,
 	[IDManager] [bigint] NULL,
-	[IDCounty] [bigint] NOT NULL,
-	[PlanYTD] [money] NULL,
-	[ActualYTD] [money] NULL
+	--[PlanYTD] [money] NULL,
+	--[ActualYTD] [money] NULL
 -- CONSTRAINT [PK_KPI11] PRIMARY KEY CLUSTERED 
 --(
 --	[ID] ASC
@@ -23,10 +22,13 @@ CREATE TABLE [dbo].[KPI11](
 ) ON [PRIMARY]
 
 go
+
 drop procedure createKPI11
 --DROP TABLE KPI11
 go
-CREATE PROCEDURE createKPI11
+CREATE PROCEDURE createKPI11(
+@userId nvarchar(1000),
+@managers nvarchar(1000), @clients nvarchar(100))
 AS
 BEGIN
 	SET NOCOUNT ON;
@@ -35,25 +37,66 @@ declare @year int, @month int
 set @year = Year(getdate())
 set @month = MONTH(getdate())
 
-delete from KPI11
+delete from KPI11  where @userId = UserId
+
+declare @managersID table(id int identity, ManagerVar int)
+insert into @managersID(ManagerVar)
+select * from String_Split(@managers,',')
+
+declare @managerVar int
+DECLARE @I INT, @maxManagers int
+select @maxManagers  = count(*) from @managersID
+print @maxManagers
+set @I=1
+
+WHILE @I <= @maxManagers
+BEGIN
+select @managerVar= ManagerVar from @managersID where id=@I
+--print @managerVar
+--select * from @managersID where id=@I
+;WITH cte_org AS (
+    SELECT 
+		a.IDAssVA,a.IDManager, a.NameAssVA , cast ('' as nvarchar(100))  Manager
+	 FROM AssVA a
+	 --inner join @managersID on a.IDManager = ManagerVar
+    WHERE idmanager =@managerVar
+    UNION ALL
+    SELECT a.IDAssVA,a.IDManager,a.NameAssVA as AssVA, c.NameAssVA as Manager
+    FROM 
+        AssVA a
+        INNER JOIN cte_org c 
+            ON a.IDManager =c.IDAssVA 
+)
 
 INSERT INTO [KPI11]
-           ([IDClient]
+           (
+		   [UserId]
            ,[IDAssVA]
            ,[IDManager]
-           ,[IDCounty]
-           ,[PlanYTD]
-           ,[ActualYTD])
-select IDClient, c.IDAssVA,m.IDManager, IDCounty,sum([Plan]) as PlanYTD,sum(Actual) as ActualYTD 
---sinto KPI11
-from vwACTPL_Ass_Clients c 
-inner join AssVA m on c.IDAssVA = m.IDManager
-where c.Year =Year and c.Month <@Month
-group by IDClient, c.IDAssVA,IDCounty, m.IDManager
+)
+SELECT @userId, IDAssVA,@managerVar  FROM cte_org
+
+Set @I= @I+1
+end
+--INSERT INTO [KPI11]
+--           (
+--		   [UserId]
+--		   ,[IDClient]
+--           ,[IDAssVA]
+--           ,[IDManager]
+--           ,[PlanYTD]
+--           ,[ActualYTD])
+--select IDClient, c.IDAssVA,m.IDManager, IDCounty,sum([Plan]) as PlanYTD,sum(Actual) as ActualYTD 
+----sinto KPI11
+--from vwACTPL_Ass_Clients c 
+--inner join AssVA m on c.IDAssVA = m.IDManager
+--where c.Year =Year and c.Month <@Month
+--group by IDClient, c.IDAssVA,IDCounty, m.IDManager
 
 
 
 END
 GO
 
-exec createKPI11
+exec createKPI11 'a','20,19',''
+select * from KPI11
